requirements.txt
pytest>=7.0.0
allure-pytest>=2.12.0

pyodbc>=4.0.32
psycopg2-binary>=2.9.0
PyYAML>=6.0

requests==2.31.0
boto3==1.34.0
google-cloud-storage==2.10.0
selenium==4.15.2
webdriver-manager==4.0.1
